# tools/api/mmc_atlas_api_client/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: MMC Atlas API Helper
  title: MMC Atlas API Helper
  version: "1.0"
servers:
- url: //localhost:8080/api
paths:
...
  /api/project/{ProjectId}/cluster/{ClusterName}/restore:
    get:
      description: Returns the restore job for the specified cluster.
      operationId: GetRestoreJob
      parameters:
      - description: The name of the cluster to restore.
        explode: false
        in: path
        name: ClusterName
        required: true
        schema:
          type: string
        style: simple
      - description: The ID of the project that contains the cluster to restore.
        explode: false
        in: path
        name: ProjectId
        required: true
        schema:
          type: string
        style: simple
      - description: Public Key
        explode: false
        in: header
        name: x-mongo-publickey
        required: true
        schema:
          default: <publicKey>
          type: string
        style: simple
      - description: Private Key
        explode: false
        in: header
        name: x-mongo-privatekey
        required: true
        schema:
          default: <privateKey>
          type: string
        style: simple
      - description: The JobID of the cluster to restore.
        explode: true
        in: query
        name: JobId
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Internal Server Error
      summary: Get Restore Job
      tags:
      - Cluster Restore
    post:
      description: Creates a restore job for the specified cluster.
      operationId: CreateRestoreJob
      parameters:
      - description: The name of the cluster to restore.
        explode: false
        in: path
        name: ClusterName
        required: true
        schema:
          type: string
        style: simple
      - description: The ID of the project that contains the cluster to restore.
        explode: false
        in: path
        name: ProjectId
        required: true
        schema:
          type: string
        style: simple
      - description: Public Key
        explode: false
        in: header
        name: x-mongo-publickey
        required: true
        schema:
          default: <publicKey>
          type: string
        style: simple
      - description: Private Key
        explode: false
        in: header
        name: x-mongo-privatekey
        required: true
        schema:
          default: <privateKey>
          type: string
        style: simple
      - description: The ID of the snapshot to restore.
        explode: true
        in: query
        name: SnapshotId
        required: false
        schema:
          type: string
        style: form
      - description: The timestamp of the oplog to restore.
        explode: true
        in: query
        name: OpLogTs
        required: false
        schema:
          type: integer
        style: form
      - description: The increment of the oplog to restore.
        explode: true
        in: query
        name: OpLogInc
        required: false
        schema:
          type: integer
        style: form
      - description: "The point-in-time to restore, specified as a Unix timestamp\
          \ in seconds."
        explode: true
        in: query
        name: PointInTimeUtcSeconds
        required: false
        schema:
          type: integer
        style: form
      - description: The name of the target cluster.
        explode: true
        in: query
        name: TargetClusterName
        required: false
        schema:
          type: string
        style: form
      - description: The ID of the project that contains the target cluster.
        explode: true
        in: query
        name: TargetProjectId
        required: false
        schema:
          type: string
        style: form
      - description: The delivery type of the restore job.
        explode: true
        in: query
        name: DeliveryType
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AtlasResponse'
          description: Internal Server Error
      tags:
      - Cluster Restore
...
components:
  schemas:
...
    AtlasResponse:
      example:
        response: "{}"
        errorMessage: errorMessage
        errorCode: errorCode
        status: status
      properties:
        errorCode:
          type: string
        errorMessage:
          type: string
        response:
          type: object
        status:
          type: string
      type: object
  securitySchemes:
    BasicAuth:
      scheme: basic
      type: http
...